fix code

+++ product/src/main/java/com/vnco/fusiontech/product/repository/CategoryRepository.java
@@ -3,5 +3,7 @@ package com.vnco.fusiontech.product.repository;
 import com.vnco.fusiontech.product.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
+public interface CategoryRepository extends JpaRepository<Category,Long> {
+
 }
